% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoTemp_foldChange.R
\name{twoTemp_foldChange}
\alias{twoTemp_foldChange}
\title{This function allows you to calculate the fold change of two different temperatures and to visualize.}
\usage{
twoTemp_foldChange(rawData_inputfilename1, rawData_inputfilename2,
  seqCol = 3, protAccCol = 10, desCol = 13, sp1ctrlCol = 16,
  sp2ctrlCol = 17, sp1lowCol = 18, sp2lowCol = 19, sp1highCol = 20,
  sp2highCol = 21, workingDir, mapDIA_flag = F, mapDIADir,
  userSystem = "Linux", defaultInput_parameter, cutoff_uFc = 1.3,
  cutoff_lFc = 0.7, cutoff_mapDIA = 0.05, outputTagS1, outputTagS2)
}
\arguments{
\item{rawDate_inputfilename1}{absolute path to your input file of temperature 1}
}
\description{
This function allows you to calculate the fold change of two different temperatures and to visualize.
}
